Move `src/facet.ts` to `src/spec/facet.ts`
