+++ src/spec/facet.ts
@@ -1,11 +1,5 @@
-import {ChannelDef, RepeatRef, SortableFieldDef} from '../fielddef';
+import {ChannelDef, SortableFieldDef} from '../fielddef';
 import {Header} from '../header';
-import {Resolve} from '../resolve';
-import {BaseSpec} from './base';
-import {FacetMapping} from './facet';
-import {GenericLayerSpec, NormalizedLayerSpec} from './layer';
-import {GenericCompositionLayout} from './toplevel';
-import {GenericUnitSpec, NormalizedUnitSpec} from './unit';
 
 export interface FacetFieldDef<F> extends SortableFieldDef<F> {
   /**
@@ -28,30 +22,4 @@ export interface FacetMapping<F> {
 
 export function isFacetFieldDef<F>(channelDef: ChannelDef<F>): channelDef is FacetFieldDef<F> {
   return !!channelDef && !!channelDef['header'];
-}
-
-export interface GenericFacetSpec<U extends GenericUnitSpec<any, any>, L extends GenericLayerSpec<any>>
-  extends BaseSpec,
-    GenericCompositionLayout {
-  /**
-   * An object that describes mappings between `row` and `column` channels and their field definitions.
-   */
-  facet: FacetMapping<string | RepeatRef>;
-
-  /**
-   * A specification of the view that gets faceted.
-   */
-  spec: L | U;
-  // TODO: replace this with GenericSpec<U> once we support all cases;
-
-  /**
-   * Scale, axis, and legend resolutions for facets.
-   */
-  resolve?: Resolve;
-}
-
-export type NormalizedFacetSpec = GenericFacetSpec<NormalizedUnitSpec, NormalizedLayerSpec>;
-
-export function isFacetSpec(spec: BaseSpec): spec is GenericFacetSpec<any, any> {
-  return spec['facet'] !== undefined;
 }
